requirements.txt
streamlit
python-docx
python-dotenv
openai
PyMuPDF
docx2txt
